refactor filer interfaces
