+++ frontend/components/DailyTabs.tsx
@@ -5,8 +5,7 @@ import Box from '@mui/material/Box';
 
 import TimeComponent from './TimeComponent';
 
-import { UniqueFilter } from '@/helpers/getData';
-import { Choice } from '@/pages';
+import { UniqueFilter } from '@/interfaces/Filters';
 
 interface TabPanelProps {
   children?: ReactNode;
