+++ frontend/interfaces/Filters.tsx
@@ -1,0 +1,3 @@
+export interface UniqueFilter {
+  [index: string]: Array<number>
+}
